requirements.txt
pandas
happybase
mysql-connector-python
pyspark
confluent-kafka